Add tool for generating single-include header
The "single_include" directory was previously managed by hand, which is
not a particularly maintainable approach long-term since it requires
manually porting of changes made to the string_view.hpp/inl file to the
joined file.

To make this simpler, a generate-single-include.py file has been added
to the repository instead.